Added check for additional existing versions locations and update of existing installs of youtube-dl

+++ install.sh
@@ -1,13 +1,16 @@
 #!/bin/sh
 
 # Install youtube-dl. Use github to make sure it's the latest version.
-if [ ! -f /usr/local/bin/youtube-dl ];
+# Check for /usr/bin/youtube-dl (from package manager install) and
+# /usr/local/bin/youtube-dl (from previous repo based install)
+if [ ! -f /usr/bin/youtube-dl -a ! -f /usr/local/bin/youtube-dl ];
 then
     echo "> Installing youtube-dl."
     sudo wget https://github.com/downloads/rg3/youtube-dl/youtube-dl -O /usr/local/bin/youtube-dl
     sudo chmod a+x /usr/local/bin/youtube-dl
 else
-    echo "> Skipping installing youtube-dl, already installed."
+    echo "> Updating youtube-dl."
+    sudo youtube-dl -U
 fi
 
 # Install dependancies
@@ -17,7 +20,7 @@ sudo apt-get install omxplayer wget libjpeg8
 # Download binary
 echo "> Downloading raspytube binary."
 sudo wget https://github.com/bbond007/raspytube/blob/master/raspytube.bin -O /usr/local/bin/raspytube.bin
-chmod a+x /usr/local/bin/raspytube.bin
+sudo chmod a+x /usr/local/bin/raspytube.bin
 
 # Launch raspytube!
 echo "> Launching raspytube!"
